Hide ProjectOperationCredentials type in GitHubTargetsParams
