+++ src/operations/review/issueRaisingReviewRouter.ts
@@ -3,6 +3,7 @@ import { failureOn, successOn } from "../../action/ActionResult";
 import { deepLink, Issue, raiseIssue } from "../../util/gitHub";
 import { GitHubRepoRef, isGitHubRepoRef } from "../common/GitHubRepoRef";
 import { EditorOrReviewerParameters } from "../common/params/BaseEditorOrReviewerParameters";
+import { TokenCredentials } from "../common/ProjectOperationCredentials";
 import { ReviewRouter } from "./reviewerToCommand";
 import { ProjectReview, ReviewComment } from "./ReviewResult";
 
@@ -17,7 +18,7 @@ export const issueRaisingReviewRouter: ReviewRouter<EditorOrReviewerParameters> 
     (pr: ProjectReview, params: EditorOrReviewerParameters, name: string) => {
         if (isGitHubRepoRef(pr.repoId)) {
             const issue = toIssue(pr, name);
-            return raiseIssue(params.targets.githubToken, pr.repoId, issue)
+            return raiseIssue((params.targets.credentials as TokenCredentials).token, pr.repoId, issue)
                 .then(ap => successOn(pr.repoId));
         } else {
             return Promise.resolve(failureOn(pr.repoId, new Error(`Not a GitHub Repo: ${JSON.stringify(pr.repoId)}`)));
